% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claeaning_rspo.R
\name{clean_rspo}
\alias{clean_rspo}
\title{Czyszczenie zrzutu z bazy RSPO}
\usage{
clean_rspo(x)
}
\arguments{
\item{x}{ramka danych będąca bazą RSPO - najlepiej wynik działania funkcji
\code{[get_raw_rspo()]}}
}
\value{
data.frame
}
\description{
Funkcja wykonująca wszystkie operacje czyszczenia zbioru RSPO po
kolei, czyli:
\itemize{
\item{\code{[df_into_vector()]}}{konwersja ramek danych na wektory}
\item{\code{[remove_columns()]}}{usuwanie zbędnych kolumn}
\item{\code{[convert_date()]}}{konwersja na daty i zmienne liczbowe}
\item{\code{[new_columns()]}}{tworzenie nowych kolumn}
}
}
