% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claeaning_rspo.R
\name{convert_date}
\alias{convert_date}
\title{Czyszczenie zrzutu z bazy RSPO}
\usage{
convert_date(x)
}
\arguments{
\item{x}{ramka danych będąca bazą RSPO - najlepiej wynik działania funkcji
\code{[get_raw_rspo()]}}
}
\value{
data.frame
}
\description{
Funkcja konwertuje wartości numeryczne, które są przechowywane
jako wartości tekstowe w niektórych kolumnach zbioru, na wartości klasy
\code{integer} oraz wartości będące datą, ale przechowywane jako tekst,
na wartości klasy \code{date}.
}
