% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claeaning_rspo.R
\name{df_into_vector}
\alias{df_into_vector}
\title{Czyszczenie zrzutu z bazy RSPO}
\usage{
df_into_vector(x, listy = FALSE)
}
\arguments{
\item{x}{ramka danych będąca bazą RSPO - najlepiej wynik działania funkcji
\code{[get_raw_rspo()]}}

\item{listy}{czy mają zostać zwrócone zmienne będące listami? Zmienne jako
listy mogą się nie zapisywać do pliku .csv}
}
\value{
data.frame
}
\description{
Funkcja spłaszcza ramkę danych konwertując kolumny zawierające
inne ramki danych lub listy do wektorów. Zarówno ramki danych, jak i listy
mają podobne struktury, gdzie istotne informacje znajdują się zazwyczaj w
tylko jednej kolumnie/elemencie listy, dlatego konwersja często polega na
wyciągnięciu tej wartości i do osobnej kolumny zbioru RSPO.
Niektóre listy są na tyle rozbudowane, że nie można w prosty sposób ich
spłaszczyć, dlatego użytkownik może za pomocą parametru \code{listy}
zdecydować czy wykluczyć ze zbioru te listy czy zostawić je w formie
niezmienionej (niespłaszczonej). W przypadku decyzji o niespłąszczaniu list
mogą wystąpić problemy z zapisywaniem zbioru z takimi kolumnami do pliku
o rozszerzeniu .csv.
}
