% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zrzut_rspo.R
\name{get_df_api}
\alias{get_df_api}
\title{Tworzenie zrzutu bazy RSPO}
\usage{
get_df_api(
  api_adress = "https://api-rspo.mein.gov.pl/api/placowki",
  page_number
)
}
\arguments{
\item{api_adress}{ciąg znaków będących linkiem do strony z API bazy RSPO}

\item{page_number}{wartość liczbowa będąca numerem strony API. API bazy RSPO
ma ograniczenie do zwracania 100 obserwacji na stronę. Ten argument pozwala
wybrać, która strona ma zostać pobrana.}
}
\value{
data.frame
}
\description{
Tworzenie zrzutu bazy RSPO
}
\details{
Funkcja zwraca przy użyciu API bazy RSPO surową ramkę danych dla
zadanego numeru strony
}
