% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zrzut_rspo.R
\name{get_raw_rspo}
\alias{get_raw_rspo}
\title{Tworzenie zrzutu bazy RSPO}
\usage{
get_raw_rspo(max_page = 999999)
}
\arguments{
\item{max_page}{liczba stron, które mają zostać pobrane z api. Domyślnie jest
to wartość 999999 (przyjęto, że jest to wystarczająco wysoka wartość, żeby
liczba rzeczywistych stron nie przekroczyła tej wartości), bo domyślnie
przyjmuje się, że użytkownik chce pobrać całą bazę i raczej nie ma wiedzy o
tym, co konkretnie jest w danych 100 obserwacjach, które można jednorazowo
pobrać przez API.}
}
\value{
data.frame
}
\description{
Tworzenie zrzutu bazy RSPO
}
\details{
Funkcja zwraca przy użyciu API bazy RSPO surową ramkę danych dla
zadanego zakresu stron (użytkownik podaje tylko maksymalną wartość zakresu
stron), czyli domyślnie dla całego zbioru.
}
