% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ibe_losy.R
\name{ibe_miejsc}
\alias{ibe_miejsc}
\title{Czyszczenie zrzutu z bazy RSPO na potrzeby projektu Losy2}
\usage{
ibe_miejsc(x)
}
\arguments{
\item{x}{ramka danych będąca wyczyszczoną bazą RSPO - najlepiej wynik
działania funkcji \code{[clean_rspo()]}}
}
\value{
data.frame
}
\description{
Zmienna \code{kl_wlk_miejsc} powstaje na podstawie kodu TERYT na
poziomie gminy (kod 6-cio lub 7-mio cyfrowy, zawierający w ostatniej cyfrze
kodu informację o rodzaju gminy (w bazie RSPO sama ta informacja
przechowywana jest dodatkowo w zmiennej \code{gminaRodzajKod})). Poniżej
opisano logikę tworzenia zmiennej na podstawie wspomnianego numeru TERYT:
\itemize{
\item{miasta na prawach powiatu}{Jako, że podział jest tworzony na poziomie
całych gmin, to nie są brane pod uwagę obszary miejskie i obszary wiejskie
w ramach gmin miejsko-wiejskich ani dzielnice ani delegatury. Miasta na
prawach powiatu to gminy miejskie albo powiaty, których 2 cyfry TERYTu
(czyli 3. i 4. wartość TERYTu) opisujące powiat są większe od 60.}
\item{miasta, wsie i obszary miejsko-wiejskie}{Pozostałe 3 kategorie
wyodrębniane są na podstawie 7. cyfry w numerze TERYT, która w rejestrze
TERC jest wartością zmiennej RODZ. Kolejne wartości podziału kodowane są
następująco: 1 = miasto, 2 = wieś, 3 = obszar miejsko-wiejski}
}
Zmienna \code{kl_wlk_miejsc_3} jest ważną zmienną w pracy ze zbiorami z
sondaży realizowanych w projekcie Losy2 w IBE, ponieważ to do jej rozkładu w
populacji absolwentów ważone są dane sondażowe.
}
\details{
Funkcja oblicza na podstawie zbioru RSPO 3 nowe kolumny na potrzeby
prac w projekcie Losy2 w Instytucie Badań Edukacyjnych. Funkcja tworzy
następujące zmienne:
\itemize{
\item{\code{kl_wlk_miejsc}}{informacja o klasie wielkości miejscowości,
w której znajduje się placówka. Przyjmuje 4 wartości: \code{mnpp} - miasta z
delegaturami oraz miasta na prawach powiatu; \code{m} - miasta powiatowe; \code{w} -
wieś; \code{mw} - inne miasta}
\item{\code{kl_wlk_miejsc_3}}{informacja o klasie wielkości miejscowości,
ale zawężona do 3 kategorii poprzez połączenie dwóch kategorii zmiennej
\code{kl_wlk_miejsc} - połączono kategorie \code{w} i \code{mw} w jedną}
\item{\code{miasto}}{zmienna binarna o wartości \code{TRUE} jeśli szkoła
znajduje w obszarze miejskim i \code{FALSE} jeśli nie znajduje się w obszarze
miejskim}
}
}
