% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claeaning_rspo.R
\name{new_columns}
\alias{new_columns}
\title{Czyszczenie zrzutu z bazy RSPO}
\usage{
new_columns(x)
}
\arguments{
\item{x}{ramka danych będąca bazą RSPO - najlepiej wynik działania funkcji
\code{[get_raw_rspo()]}}
}
\value{
data.frame
}
\description{
Funkcja tworzy nowe zmienne na podstwie zmiennych z bazy RSPO, o
ile zmienne składowe w niej występują.
Tworzone są zmienne:
\itemize{
\item{\code{adresSzkola}}{pełny adres szkoły}
\item{\code{adresDoKorespondencjiSzkola}}{pełny adres korespondencyjny
szkoły}
\item{\code{dyrektor}}(pełne imię i nazwisko dyrektora szkoły)
}
}
