% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claeaning_rspo.R
\name{remove_columns}
\alias{remove_columns}
\title{Czyszczenie zrzutu z bazy RSPO}
\usage{
remove_columns(x)
}
\arguments{
\item{x}{ramka danych będąca bazą RSPO - najlepiej wynik działania funkcji
\code{[get_raw_rspo()]}}
}
\value{
data.frame
}
\description{
Funkcja usuwa niepotrzebne kolumny ze zbioru RSPO, jeśli takowe
występują. Aby funkcja spełniała swoje założenia najlepiej, aby zbiór
przekazywany do argumentu \code{x} był wynikiem działania funkcji
\code{[get_raw_rspo()]}.
}
